requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.36
matplotlib==3.8.2
seaborn==0.13.1

# Optimization & Scientific Computing
scipy==1.12.0
scikit-learn==1.4.0

# LLM Integration
requests==2.31.0
huggingface-hub==0.20.3

# Additional utilities
python-dotenv==1.0.0
plotly==5.18.0